# src/main/java/com/supera/accessmanager/dto/auth/LoginRequest.java
package com.supera.accessmanager.dto.auth;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record LoginRequest(
        @Email @NotBlank String email,
        @NotBlank String senha
) {}