# src/main/java/com/supera/accessmanager/exception/ErrorResponse.java
package com.supera.accessmanager.exception;

import java.time.LocalDateTime;

public record ErrorResponse(
        LocalDateTime timestamp,
        int status,
        String erro,
        String mensagem,
        String path
) {}